+++ rai-sdk/src/test/java/com/relationalai/ExecuteAsyncTest.java
@@ -36,7 +36,7 @@ public class ExecuteAsyncTest extends UnitTest {
 
         var query = "x, x^2, x^3, x^4 from x in {1; 2; 3; 4; 5}";
 
-        var rsp = client.executeAsyncWait(databaseName, engineName, query, true);
+        var rsp = client.execute(databaseName, engineName, query, true);
 
         var results = new ArrayList<ArrowRelation> () {
             {