+++ rai-sdk/src/test/java/com/relationalai/LoadCsvTest.java
@@ -47,7 +47,7 @@ public class LoadCsvTest extends UnitTest {
         assertEquals(0, loadRsp.output.length);
         assertEquals(0, loadRsp.problems.length);
 
-        var rsp = client.execute(databaseName, engineName, "def output = sample");
+        var rsp = client.executeV1(databaseName, engineName, "def output = sample");
 
         Relation rel;
 
@@ -103,7 +103,7 @@ public class LoadCsvTest extends UnitTest {
         assertEquals(0, loadRsp.output.length);
         assertEquals(0, loadRsp.problems.length);
 
-        var rsp = client.execute(databaseName, engineName, "def output = sample_no_header");
+        var rsp = client.executeV1(databaseName, engineName, "def output = sample_no_header");
 
         Relation rel;
 
@@ -160,7 +160,7 @@ public class LoadCsvTest extends UnitTest {
         assertEquals(0, loadRsp.output.length);
         assertEquals(0, loadRsp.problems.length);
 
-        var rsp = client.execute(databaseName, engineName, "def output = sample_alt_syntax");
+        var rsp = client.executeV1(databaseName, engineName, "def output = sample_alt_syntax");
 
         Relation rel;
 
@@ -214,7 +214,7 @@ public class LoadCsvTest extends UnitTest {
         assertEquals(0, loadRsp.output.length);
         assertEquals(0, loadRsp.problems.length);
 
-        var rsp = client.execute(databaseName, engineName, "def output = sample");
+        var rsp = client.executeV1(databaseName, engineName, "def output = sample");
 
         Relation rel;