+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Changelog
+
+## v0.4.0-alpha
+* Renamed:
+  - `execute` to `executeV1`.
+  - `executeAsyncWait` to `execute`.
+
 ## v0.3.0-alpha
 * Added `CancelTransaction` feature.